Adding Resx configuration to the vcxproj file.
In my project group we are using CMake to generate c++/cli winform
projects and I noticed the work done in "VS: Add Windows Forms Support"
was in the right direction for solving some of the problems we were
facing.

The changes as submitted was breaking some functionality in our
projects, so I made some changes that fixes our problems and I believe
that it will also work for others.

1) Resx files did not link correctly with the winform h-file so I added
the Resx configuration to the vcxproj file.

2) I removed the functionality for setting <CLRSupport> true for the
project - based on if an resx-file is pressent. This is preventing us
from using native cpp code.
Also this do not address that some projects will need to set other
options like clr:pure, clr:safe.
This could be implemented as a cmake option, so it is possible to
specify exactly what is needed.
Existing VSWindowsFormsResx Test project is updated so it will be
working with my changes.