cmTestGenerator: Evaluate generator expressions in test properties
This is useful for something like:

    set_tests_properties(mytest
      PROPERTIES
        WORKING_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}/$<CONFIGURATION>")

The common thought is that getting CTEST_CONFIGURATION_TYPE to be
expanded when setting the property is the way, but this variable is not
normalized. On Windows, using the variable in `ctest -R rELEASE` will
likely work, but with Xcode, the path does not exist due to case
sensitivity.