[CheckStructHasMember] Add support for C++
If headers required to check if a struct has a member can be compiled
with C++ compiler only, the check will fail, because the C compiler
will fail.
As a consequence, the result variable is set to false, even if the
struct has that particular member.

The CHECK_C_STRUCT_HAS_MEMBER now accepts a new optional argument
LANGUAGE that allows one to explicitly set the compiler to use.
The new signature is therefore:

  CHECK_STRUCT_HAS_MEMBER (<struct> <member> <header> <variable>
                           [LANGUAGE <language>])