Update cmTarget SourceEntries while tracing dependencies
Do not make a copy in cmGeneratorTarget and update that one.
We must update the original so that cmTarget::GetSourceDepends
returns the proper information after tracing dependencies.
This is needed for the VS 10 generator (but not VS >= 11) to
mark external objects as <None> instead of <Object> if they
are also the output of a custom command in the same target.