cmQtAutogen: Gather tool arguments after creating generator targets.
One of the side-effects in cmGlobalGenerator::CreateGeneratorTargets
is that compile definitions from the makefile are appended to the
target. Compile definitions are read from the target directly for
generating the buildsystem, and generating the Autogen info too
early caused incomplete compile definitions to be used (#14535).

The compile definitions related code will be moved out of
CreateGeneratorTargets in a followup commit. However this change
also has the independent benefit that it causes the
GetLinkInterfaceDependentStringProperty method to be called in
SetupAutoGenerateTarget, after the creation of generator targets. In
CMake 4.0, the GetLinkInterfaceDependentStringProperty will move to
the cmGeneratorTarget class, and this is a necessary prerequisite to
that.