CMakeParseArguments: Do not skip empty arguments
Current implementation of cmake_parse_arguments skips empty arguments.

For example:

  cmake_parse_arguments("OPTION" "SINGLE" "MULTI" ${ARGN})

with ARGN that is set using something like:

  <args> SINGLE "" <more args>  # (<args>;SINGLE;;<more args>)
  <args> SINGLE ""  # (<args>;SINGLE;)

it will fail in 4 different ways:

1) because the "foreach" skips empty arguments.
2) because list(APPEND <list> ${currentArg}) will fail to append an
   empty element to the list if ${currentArg} is an empty string.
3) because set(${prefix}_${arg_name} ${${prefix}_${arg_name}} PARENT_SCOPE)
   will not set an empty variable to the parent scope if
   ${${prefix}_${arg_name}} is an empty string.
4) because empty arguments are not passed to cmake_parse_arguments
   if not quoted, i.e. if ARGN="a;;b"
   * ${ARGN} will pass the arguments "a", "b"
   * "${ARGN}" will pass the arguments "a", "", "b".

This patch fixes the first behaviour using foreach(IN LISTS) version,
the second and the third by enclosing args in quotes where required, and
finally fixes the documentation in order to suggest to use quoted
"${ARGN}" in order not to miss the empty arguments at the end of the
list.